+++ BunnymodXT/wrappers/triangle_wrappers.cpp
